rename : 예약 정보 도메인 패키징 수정
