+++ app/src/main/java/woowacourse/movie/domain/model/booking/Booking.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.booking
 
 import java.io.Serializable
 import java.time.LocalDate
