+++ app/src/main/java/woowacourse/movie/domain/model/booking/DayType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.booking
 
 import java.time.DayOfWeek
 import java.time.LocalDate
