+++ app/src/main/java/woowacourse/movie/domain/model/booking/TicketType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.booking
 
 enum class TicketType(val price: Int) {
     GENERAL(13000),
