+++ app/src/test/java/woowacourse/movie/domain/PeopleCountTest.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.domain
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import woowacourse.movie.domain.model.PeopleCount
+import woowacourse.movie.domain.model.booking.PeopleCount
 
 class PeopleCountTest {
     @Test
