+++ app/src/test/java/woowacourse/movie/domain/ScreeningTimeTest.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.domain
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import woowacourse.movie.domain.model.ScreeningTime
+import woowacourse.movie.domain.model.booking.ScreeningTime
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
